requirements.txt
PyYAML>=6.0
Pillow>=10.0.0
safetensors>=0.4.3
transformers>=4.46.0
torch>=2.2.0
